functions/yahoo/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-storage-blob==12.12.0
azure-storage-file-datalake
azure-identity
azure-keyvault-secrets
azure-storage-queue
datetime
pandas
numpy
pyarrow
yfinance==0.2.17
yahooquery
requests
lxml
asyncio